Move webpack back to the global level.
