+++ package.json
@@ -79,62 +79,65 @@
 		"focus-trap-react": "^10.0.0"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.14.8",
-		"@babel/plugin-transform-runtime": "^7.10.4",
-		"@babel/preset-react": "^7.10.4",
-		"@svgr/webpack": "^5.4.0",
-		"@testing-library/react": "^10.4.9",
-		"@wordpress/babel-preset-default": "^4.17.0",
-		"@wordpress/browserslist-config": "^5.40.0",
-		"@wordpress/eslint-plugin": "^7.1.0",
-		"@wordpress/prettier-config": "^2.0",
-		"@wordpress/stylelint-config": "^21.0.0",
+		"@babel/core": "^7.26.10",
+		"@babel/plugin-transform-runtime": "^7.26.10",
+		"@babel/preset-react": "^7.26.3",
+		"@svgr/webpack": "^5.5.0",
+		"@wordpress/babel-preset-default": "^4.20.0",
+		"@wordpress/browserslist-config": "^5.41.0",
+		"@wordpress/eslint-plugin": "^7.4.0",
+		"@wordpress/prettier-config": "^2.25.13",
+		"@wordpress/stylelint-config": "^21.41.0",
 		"archiver": "^5.0.2",
-		"autoprefixer": "^9.8.5",
-		"babel-loader": "^8.1.0",
+		"autoprefixer": "^9.8.8",
+		"babel-loader": "^8.4.1",
+		"circular-dependency-plugin": "^5.2.2",
+		"copy-webpack-plugin": "^6.4.1",
+		"create-file-webpack": "^1.0.2",
 		"cross-env": "^7.0.3",
 		"css-loader": "^3.6.0",
-		"cssnano": "^4.1.10",
+		"cssnano": "^4.1.11",
 		"eslint": "^7.32.0",
 		"eslint-import-resolver-alias": "^1.1.2",
 		"eslint-plugin-filenames": "^1.3.2",
-		"eslint-plugin-import": "^2.25.3",
+		"eslint-plugin-import": "^2.31.0",
 		"eslint-plugin-jest": "^22.21.0",
 		"eslint-plugin-jsdoc": "^30.7.13",
 		"eslint-plugin-lodash": "^7.4.0",
-		"eslint-plugin-react-hooks": "^4.2.0",
+		"eslint-plugin-react-hooks": "^4.6.2",
 		"eslint-plugin-sitekit": "file:packages/eslint-plugin",
-		"eslint-scope": "^5.1.0",
-		"eslint-webpack-plugin": "^2.5.1",
-		"file-loader": "^6.1.1",
-		"fs-extra": "^9.0.1",
+		"eslint-scope": "^5.1.1",
+		"eslint-webpack-plugin": "^2.7.0",
+		"fs-extra": "^9.1.0",
 		"git-repo-info": "^2.1.1",
 		"glob-to-regexp": "^0.4.1",
 		"husky": "^3.1.0",
-		"lint-staged": "^10.2.11",
-		"mini-css-extract-plugin": "^1.2.1",
+		"lint-staged": "^10.5.4",
+		"mini-css-extract-plugin": "^1.6.2",
 		"natives": "^1.1.6",
 		"patch-package": "^8.0.0",
-		"postcss": "^8.4.31",
-		"postcss-import-url": "^7.0.0",
-		"postcss-loader": "^4.0.0",
-		"postcss-preset-env": "^9.5.11",
-		"prettier": "npm:wp-prettier@2.6.2",
-		"rimraf": "^3.0.2",
+		"postcss": "^8.5.3",
+		"postcss-import-url": "^7.2.0",
+		"postcss-loader": "^4.3.0",
+		"postcss-preset-env": "^9.6.0",
+		"prettier": "npm:wp-prettier@^2.6.2",
+		"rimraf": "^6.0.1",
 		"sanitize-filename": "^1.6.3",
 		"sass": "^1.30.0",
-		"sass-loader": "^9.0.2",
+		"sass-loader": "^10.5.2",
 		"semver-compare": "^1.0.0",
 		"semver-regex": "^3.1.4",
-		"stylelint": "^14.11.0",
+		"stylelint": "^14.16.1",
 		"stylelint-config-standard-scss": "^5.0.0",
-		"stylelint-declaration-strict-value": "^1.9.0",
-		"stylelint-order": "^5.0.0",
-		"stylelint-scss": "^4.3.0",
-		"terser-webpack-plugin": "^3.0.6",
-		"webpack": "^4.43.0",
+		"stylelint-declaration-strict-value": "^1.10.11",
+		"stylelint-order": "^6.0.4",
+		"stylelint-scss": "^4.7.0",
+		"terser-webpack-plugin": "^3.1.0",
+		"webpack": "^4.47.0",
+		"webpack-bundle-analyzer": "^4.10.2",
 		"webpack-cli": "^3.3.12",
-		"webpack-stream": "^5.2.1"
+		"webpack-manifest-plugin": "^2.2.0",
+		"webpackbar": "^4.0.0"
 	},
 	"workspaces": [
 		"assets",
